Use a builder for workflows
