+++ pkg/workflow/executor.go
@@ -1,0 +1,46 @@
+package workflow
+
+import (
+	"fmt"
+
+	"github.com/Azure/Orkestra/pkg/utils"
+	v1alpha12 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
+	corev1 "k8s.io/api/core/v1"
+)
+
+const (
+	// The set of executor actions which can be performed on a helmrelease object
+	Install ExecutorAction = "install"
+	Delete  ExecutorAction = "delete"
+)
+
+// ExecutorAction defines the set of executor actions which can be performed on a helmrelease object
+type ExecutorAction string
+
+func defaultExecutor(tplName string, action ExecutorAction) v1alpha12.Template {
+	executorArgs := []string{"--spec", "{{inputs.parameters.helmrelease}}", "--action", string(action), "--timeout", "{{inputs.parameters.timeout}}", "--interval", "10s"}
+	return v1alpha12.Template{
+		Name:               tplName,
+		ServiceAccountName: workflowServiceAccountName(),
+		Inputs: v1alpha12.Inputs{
+			Parameters: []v1alpha12.Parameter{
+				{
+					Name: HelmReleaseArg,
+				},
+				{
+					Name:    TimeoutArg,
+					Default: utils.ToStrPtr(DefaultTimeout),
+				},
+			},
+		},
+		Executor: &v1alpha12.ExecutorConfig{
+			ServiceAccountName: workflowServiceAccountName(),
+		},
+		Outputs: v1alpha12.Outputs{},
+		Container: &corev1.Container{
+			Name:  ExecutorName,
+			Image: fmt.Sprintf("%s:%s", ExecutorImage, ExecutorImageTag),
+			Args:  executorArgs,
+		},
+	}
+}
