+++ pkg/workflow/workflow.go
@@ -2,6 +2,8 @@ package workflow
 
 import (
 	"context"
+	"fmt"
+
 	v1alpha12 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -10,52 +12,120 @@ import (
 	"github.com/go-logr/logr"
 )
 
+type EngineType string
+
 const (
-	Project        = "orkestra"
-	OwnershipLabel = "owner"
-	HeritageLabel  = "heritage"
+	Forward EngineType = "forward"
+	Reverse            = "reverse"
 )
+
+var _ = ForwardEngine{}
+var _ = ReverseEngine{}
 
 type Engine interface {
 	// Generate the object required by the workflow engine
-	Generate(ctx context.Context, l logr.Logger, g *v1alpha1.ApplicationGroup) error
+	Generate(ctx context.Context) error
 	// Submit the object required by the workflow engine generated by the Generate method
-	Submit(ctx context.Context, l logr.Logger, g *v1alpha1.ApplicationGroup) error
+	Submit(ctx context.Context) error
+
+	GetLogger() logr.Logger
+}
+
+type EngineOptions struct {
+	parallelism *int64
+	stagingRepo string
+}
+
+type EngineBuilder struct {
+	client     client.Client
+	logger     logr.Logger
+	engineType EngineType
+	options    EngineOptions
+
+	nodes           map[string]v1alpha12.NodeStatus
+	forwardWorkflow *v1alpha12.Workflow
+	appGroup        *v1alpha1.ApplicationGroup
 }
 
 type ForwardEngine struct {
 	client.Client
-	stagingRepoURL string
-	parallelism    *int64
+	logr.Logger
+	EngineOptions
 
 	workflow *v1alpha12.Workflow
+	appGroup *v1alpha1.ApplicationGroup
 }
 
 type ReverseEngine struct {
 	client.Client
-	stagingRepoURL string
-	parallelism    *int64
+	logr.Logger
+	EngineOptions
 
-	nodes map[string]v1alpha12.NodeStatus
+	nodes           map[string]v1alpha12.NodeStatus
 	forwardWorkflow *v1alpha12.Workflow
 	reverseWorkflow *v1alpha12.Workflow
 }
 
-func WorkflowFactory(workflowType string, client client.Client, stagingRepoURL string, workflowParallelism *int64) Engine {
-	switch workflowType {
-	case "forward":
+func NewEngineBuilder(client client.Client, logger logr.Logger) *EngineBuilder {
+	return &EngineBuilder{
+		client:  client,
+		logger:  logger,
+		options: EngineOptions{},
+	}
+}
+
+func (builder *EngineBuilder) Forward(appGroup *v1alpha1.ApplicationGroup) *EngineBuilder {
+	builder.engineType = Forward
+	builder.appGroup = appGroup
+	return builder
+}
+
+func (builder *EngineBuilder) Reverse(forwardWorkflow *v1alpha12.Workflow, nodes map[string]v1alpha12.NodeStatus) *EngineBuilder {
+	builder.engineType = Reverse
+	builder.forwardWorkflow = forwardWorkflow
+	builder.nodes = nodes
+	return builder
+}
+
+func (builder *EngineBuilder) WithParallelism(numNodes int64) *EngineBuilder {
+	builder.options.parallelism = &numNodes
+	return builder
+}
+
+func (builder *EngineBuilder) WithStagingRepo(stagingUrl string) *EngineBuilder {
+	builder.options.stagingRepo = stagingUrl
+	return builder
+}
+
+func (builder *EngineBuilder) Build() (Engine, error) {
+	switch builder.engineType {
+	case Forward:
 		return ForwardEngine{
-			Client: client,
-			stagingRepoURL: stagingRepoURL,
-			parallelism: workflowParallelism,
-		}
-	case "backward":
+			Client:        builder.client,
+			Logger:        builder.logger,
+			EngineOptions: builder.options,
+			appGroup:      builder.appGroup,
+		}, nil
+	case Reverse:
 		return ReverseEngine{
-			Client: client,
-			stagingRepoURL: stagingRepoURL,
-			parallelism: workflowParallelism,
-			nodes
-		}
+			Client:          builder.client,
+			Logger:          builder.logger,
+			EngineOptions:   builder.options,
+			forwardWorkflow: builder.forwardWorkflow,
+			nodes:           builder.nodes,
+		}, nil
+	}
+	return nil, fmt.Errorf("failed to build engine becasue type wasn't specified")
+}
+
+func Run(ctx context.Context, engine Engine) error {
+	if err := engine.Generate(ctx); err != nil {
+		engine.GetLogger().Error(err, "engine failed to generate workflow")
+		return fmt.Errorf("failed to generate workflow : %w", err)
+	}
+	if err := engine.Submit(ctx); err != nil {
+		engine.GetLogger().Error(err, "engine failed to submit reverse workflow")
+		return err
 	}
 	return nil
 }
